app/requirements.txt
fastapi
uvicorn
pydantic
python-socketio
asyncio

